requirements.txt
streamlit==1.28.1
llama-index==0.9.48
llama-index-llms-groq==0.1.3
llama-index-embeddings-huggingface==0.1.7
python-dotenv==1.0.0
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
pypdf==3.17.1
chromadb==0.4.21
